Adjust Protocols Node
1) The childern of the Node Protocols don't have to appear in order, see
DIN 15800:2020-07, section 14 "Supported Protocol Collect". Therefore,
the enclosing group was changed from xs:sequency to xs:all.
If a protocol node should be able to occur multiple times, this has to
be changed, e.g. to xs:choice.

2) The RDM node is renamed to FTRDM. The name RDM is a typo, but kept
around as substitution group since the DIN 15800:2020-07 is ambiguous
on page 38.
See
https://gdtf-share.com/forum/index.php?/topic/444-rdm-node-has-inconsistent-name-in-din-standard/
and the corresponding GitHub issue
https://github.com/mvrdevelopment/spec/issues/55