Rework dataversion Type
The DataVersion is defined by DIN 15800:2020-07 as "Uint.Uint" where
each Uint has 1 byte. This is now checked with RegEx. Somewhat
arbitrarily, I chose to allow a leading plus sign and zeros for the
first Uint. The second Uint cannot contain a leading plus sign or
zeros as this would break the relationship to floating point numbers.
However, the second Uint is allowed to consist of arbitrarily many
zeros to represent zero.

The valid DataVersion is restricted to the range 1.0 to 1.1. Together
with the RegEx constraint this only allows the versions 1.0 and 1.1
in this schema.
At the time of this commit, the schema is being updated to version 1.1
(DIN 15800:2020-07) which should be backwards compatible to 1.0, but
not older versions. Therefore, 0.x versions are prohibited.