Remove Unique ChannelFunction in LogicalChannel
The uniqueness constraint led to issues when different ModeMasters were
used for otherwise the same ChannelFunctions inside a LogicalChannel.

The idea was probably to test that ChannelFunctions don't overlap and
therefore only one is active at a time. The complex interaction between
ModeMasters and DMXFrom and other nodes is rather complex though and
probably not even testable in XSD 1.0. Therefore I removed this test.