Remove Irritating Default Attribute Values
All Defaults, where there is a difference in meaning between not setting
the field and the default value were removed.

Example:
Just because the Color of a Slot is not set does not mean it is the
default value. It may just not be known at creation of the fixture.
A visualizer should of course still render it with the default color
value, but a GDTF File Viewer should show the user that this field is
not set, so that the user can understand the measurement may not have
been available.

Example:
PhysicalFrom may not be available during creation of the GDTF file.
Visualizers should assume the default value of 1, but a user may want
to see whether this field is set or not to find out whether the provided
value can be assumed to be accurate or just the the default value.

Example:
Some Fixtures may have an actual Weight of 0, for example a Plug-In on
a media server. This needs to be distinguishable from "Weight unknown".

Example:
LegHeight without Value means "Legs, length unknown". LegHeight with
Value 0 means "Legs, 0 length" which is basically the same as no
LegHeight and "No Legs".

If in question, no default value was assigned.