Rework all xs:element to xs:complexType
Previously, a file like this would validate:
```
<?xml version="1.0" encoding="UTF-8" standalone="no" ?>
<FTPreset/>
```
which clearly is not a valid GDTF file.

This is because all elements were declared globally. To remedy this,
all global elements were converted to complexTypes and these types
referenced in the locally declared elements.

This is known as "Venetian Blind" instead of "Salami Slice" style
(http://www.xfront.com/GlobalVersusLocal.pdf).

Unfortunately, since uniqueness constraints can only occur inside
elements, this means uniqueness constraints are separated from the
definition of content and attributes of the type.